requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
configparser==5.2.0
discord.py==1.7.3
distlib==0.3.4
filelock==3.7.1
idna==3.3
multidict==6.0.2
pbr==5.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
six==1.16.0
stevedore==3.5.0
typing_extensions==4.2.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
yarl==1.7.2
